kafka: remove undefined behaviour when invalid data is being sent to variable-length uint deserializer

+++ source/extensions/filters/network/kafka/serialization.h
@@ -190,7 +190,10 @@ public:
       processed++;
 
       // Put the 7 bits where they should have been.
-      result_ |= ((el & 0x7f) << offset_);
+      // Impl note: the cast is done to avoid undefined behaviour when offset_ >= 28 and the highest
+      // bit in el is set (what might occur when garbage data is being sent and then it would be
+      // treated as negative signed integer).
+      result_ |= ((static_cast<uint32_t>(el) & 0x7f) << offset_);
       if ((el & 0x80) == 0) {
         // If this was the last byte to process (what is marked by unset highest bit), we are done.
         ready_ = true;
